requirements.txt
Django
django-storages
django-heroku
djangorestframework
django-rest-swagger
markdown
gunicorn
pytz
coreapi
